requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
google-generativeai==0.8.3
qdrant-client==1.12.0
psycopg2==2.9.9
sqlalchemy==2.0.35
httpx==0.27.2
python-multipart==0.0.12
markdown==3.7
beautifulsoup4==4.12.3
langchain-text-splitters==0.3.2
